+++ cedar-policy-core/src/evaluator.rs
@@ -276,6 +276,7 @@ impl<'e> Evaluator<'e> {
     /// `partial_interpret()` -- ie, so we can make sure the source locations of
     /// all errors are set properly before returning them from
     /// `partial_interpret()`.
+    #[allow(clippy::cognitive_complexity)]
     fn partial_interpret_internal(&self, expr: &Expr, slots: &SlotEnv) -> Result<PartialValue> {
         let loc = expr.source_loc(); // the `loc` describing the location of the entire expression
         match expr.expr_kind() {
@@ -656,7 +657,7 @@ impl<'e> Evaluator<'e> {
                     PartialValue::Value(v) => {
                         Ok((pattern.wildcard_match(v.get_as_string()?)).into())
                     }
-                    PartialValue::Residual(r) => Ok(Expr::like(r, pattern.iter().cloned()).into()),
+                    PartialValue::Residual(r) => Ok(Expr::like(r, pattern.clone()).into()),
                 }
             }
             ExprKind::Is { expr, entity_type } => {
@@ -2203,7 +2204,7 @@ pub mod test {
                     ("foo".into(), Expr::val(2)),
                     (
                         "bar".into(),
-                        Expr::set(vec!(Expr::val(3), Expr::val(33), Expr::val(333)))
+                        Expr::set(vec![Expr::val(3), Expr::val(33), Expr::val(333)])
                     )
                 ])
                 .unwrap(),
@@ -2430,7 +2431,7 @@ pub mod test {
             .collect::<HashMap<SmolStr, _>>();
         let entity = Entity::new(
             r#"Foo::"bar""#.parse().unwrap(),
-            attrs.clone(),
+            attrs,
             HashSet::new(),
             [],
             Extensions::none(),
@@ -3386,7 +3387,7 @@ pub mod test {
                     "decimal".parse().unwrap(),
                     vec![Value::from("3.0").into()]))),
             Err(EvaluationError::TypeError(TypeError { expected, actual, advice, .. })) => {
-                assert_eq!(expected, nonempty![Type::Extension { name: datetime_constructor.clone() }, Type::Extension { name: duration_constructor.clone() }]);
+                assert_eq!(expected, nonempty![Type::Extension { name: datetime_constructor }, Type::Extension { name: duration_constructor }]);
                 assert_eq!(actual, Type::Extension { name: "decimal".parse().unwrap() });
                 assert_eq!(advice, Some("Only extension types `datetime` and `duration` support operator overloading".into()));
         });
@@ -4233,7 +4234,7 @@ pub mod test {
         );
         // type error
         assert_matches!(
-            eval.interpret_inline_policy(&Expr::like(Expr::val(354), vec![])),
+            eval.interpret_inline_policy(&Expr::like(Expr::val(354), Pattern::from(vec![]))),
             Err(EvaluationError::TypeError(TypeError { expected, actual, advice, .. })) => {
                 assert_eq!(expected, nonempty![Type::String]);
                 assert_eq!(actual, Type::Long);
@@ -4256,7 +4257,7 @@ pub mod test {
         assert_eq!(
             eval.interpret_inline_policy(&Expr::like(
                 Expr::val("*"),
-                vec![PatternElem::Char('\u{0000}')]
+                Pattern::from(vec![PatternElem::Char('\u{0000}')])
             )),
             Ok(Value::from(false))
         );
@@ -4883,12 +4884,12 @@ pub mod test {
         ));
         assert_restricted_expression_error(Expr::like(
             Expr::val("abcdefg12"),
-            vec![
+            Pattern::from(vec![
                 PatternElem::Char('a'),
                 PatternElem::Char('b'),
                 PatternElem::Char('c'),
                 PatternElem::Wildcard,
-            ],
+            ]),
         ));
         assert_matches!(
             evaluator.partial_interpret(
@@ -5947,7 +5948,10 @@ pub mod test {
         let es = Entities::new();
         let eval = Evaluator::new(empty_request(), &es, Extensions::none());
 
-        let e = Expr::like(Expr::unknown(Unknown::new_untyped("a")), []);
+        let e = Expr::like(
+            Expr::unknown(Unknown::new_untyped("a")),
+            Pattern::from(vec![]),
+        );
 
         let r = eval.partial_interpret(&e, &HashMap::new()).unwrap();